+++ src/Bundle/WorkflowBundle/Entity/Definition/State.php
@@ -168,7 +168,7 @@ class State
 
         $this->workflow = $workflow;
 
-        if ($this->workflow) {
+        if ($this->workflow !== null) {
             $this->workflow->addState($this);
         }
 
@@ -338,7 +338,7 @@ class State
      */
     public function isDefault()
     {
-        if (isset($this->workflow)) {
+        if ($this->workflow !== null) {
             return $this === $this->workflow->getDefault();
         }
 
@@ -375,20 +375,15 @@ class State
             // NOTE: This also means that all the changes to the entity that is removed from
             // the collection wont be recorded by doctrine anymore.
 
-            if ($found = $uow->tryGetById([$permission->getGroup(), $this->getId()], \get_class($permission))) {
-                if ($found !== $permission && ($found->getState() === null || $found->getState() === $this)) {
-                    $this->permissions->removeElement($permission);
-                    $this->permissions->add($found);
-
-                    if ($uow->isInIdentityMap($permission)) {
-                        $uow->detach($permission);
-                    }
-
-                    $found->setState($this);
-                    $found->setMask($permission->getMask());
-
-                    $uow->persist($found);
+            if (($found = $uow->tryGetById([$permission->getGroup(), $this->getId()], \get_class($permission))) && ($found !== $permission && ($found->getState() === null || $found->getState() === $this))) {
+                $this->permissions->removeElement($permission);
+                $this->permissions->add($found);
+                if ($uow->isInIdentityMap($permission)) {
+                    $uow->detach($permission);
                 }
+                $found->setState($this);
+                $found->setMask($permission->getMask());
+                $uow->persist($found);
             }
         }
     }
